requirements.txt
flask
flask-mysql-connector

